ui in separate file, train/vl splits done
